+++ src/img/zimgregioniterator.h
@@ -100,7 +100,7 @@ public:
   }
 
   // go to idx of this region
-  __forceinline void goToIndex(int64_t idx)
+  __forceinline void goToIndex(size_t idx)
   {
     m_idx = idx;
     m_coord = ZImg::indexToCoord(idx, m_regionInfo);
@@ -115,7 +115,7 @@ public:
 
   // return index of current voxel of this region
   // negative index means before the region
-  __forceinline int64_t index() const
+  __forceinline size_t index() const
   { return m_idx; }
 
   template<typename OtherTImg, typename OtherTVoxel>
@@ -231,9 +231,9 @@ private:
   TImg* m_img;
   ZImgRegion m_region;
   ZImgInfo m_regionInfo;
-  int64_t m_endIdx;
+  size_t m_endIdx;
   // dynamic info of current voxel
-  int64_t m_idx;  // current voxel idx of region
+  size_t m_idx;  // current voxel idx of region
   ZVoxelCoordinate m_coord; // current voxel coord of img
 
   TVoxel* m_centerVoxelPtr; // current voxel address
